+++ CHANGELOG.md
@@ -5,8 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased - 4.4.1]
+### Added
+- intofy watches on:
+  - `charge_control_end_threshold`
+  - `panel_od`
+  - `gpu_mux_mode`
+  - `platform_profile`
+  - These allow for updating any associated config and sending dbus notifications.
 ### Changed
 - Use loops to ensure that mutex is gained for LED changes.
+### Breaking
+- DBUS: all charge control methods renamed:
+  - `ChargeControlEndThreshold`
+  - `SetChargeControlEndThreshold`
+  - `NotifyChargeControlEndThreshold`
 
 ## [v4.4.0] - 2022-08-29
 ### Added